force to array and concat to create temp array
should not write to a chef node without defining its precedence level.
Writing to the node can be avoided by using a temporary array variable.
It is more robust to force the networks node to an array and then
concatenate it, in case it was an array to begin with, otherwise you
would get a nested array. This temp var can then be joined to produce
the string variable networks.